requirements.txt
nextcord
google-genai